Check if the file exists before deleting it

+++ includes/admin.php
@@ -106,7 +106,9 @@ class Hintr_Admin {
   public function delete_json($post_type = null) : void
   {
     if (isset($post_type) && $post_type) {
-      unlink($this->plugin_uploads_path . $post_type . '.json');
+      if (file_exists($this->plugin_uploads_path . $post_type . '.json')) {
+        unlink($this->plugin_uploads_path . $post_type . '.json');
+      }
     } else {
       $files = glob($this->plugin_uploads_path . '*.json');
       foreach ($files as $file) {
